refactor: Move FluidInteractionHelper out of data/ directory Doesn't really make sense to have this on data directory
